Apuntes updatee

+++ apps/Apuntes/views.py
@@ -4,6 +4,7 @@ from django.views.generic.edit import CreateView, UpdateView
 from django.contrib.auth.mixins import LoginRequiredMixin
 from .forms import ApunteCreateForm
 from django.http import  HttpResponseRedirect
+from braces.views import UserPassesTestMixin
 # Create your views here.
 
 def apuntes(request):
@@ -14,16 +15,16 @@ def apuntes(request):
 
 def apuntes_detail(request, pk):
     apunte = get_object_or_404(Apunte, pk = pk)
-    return render(request, 'Apuntes/apuntes_detail.html', {'apunte':apunte})
+    context = {'apunte':apunte}
+    context["own_apunte"]= apunte.autor==request.user
+    return render(request, 'Apuntes/apuntes_detail.html',context)
+
+
 
 def apuntes_remove(request, pk):
     apunte = get_object_or_404(Apunte, pk=pk)
-    apunte.delete()
-    return redirect('apuntes:apuntes')
-
-
-def apuntes_remove(request, pk):
-    apunte = get_object_or_404(Apunte, pk=pk)
+    if request.user != apunte.autor : 
+        raise Http404("No eres el creador")
     apunte.delete()
     return redirect('apuntes:apuntes')
  
@@ -41,3 +42,12 @@ class ApunteCreateView(LoginRequiredMixin, CreateView):
         kwargs = super(ApunteCreateView, self).get_form_kwargs(*args, **kwargs)
         kwargs['user'] = self.request.user
         return kwargs
+
+
+class ApunteUpdateView(UserPassesTestMixin, UpdateView):
+    model = Apunte
+    fields = ['nombre', 'likes', 'paginas', 'documento', 'autor' , 'tamaño', 'thumbnail', 'categoria' , 'tipo']
+    template_name = 'Apuntes/forms/apuntes_update.html'
+
+    def test_func(self, user):
+        return user == self.get_object().autor
